Update local setup instructions to use `composer install`
Now that this codebase includes a Drupal installation, complete
with customized composer.json and composer.lock files, we can
switch from `composer create-project`, which creates a new blank
project, to `composer install`, which installs the dependencies
for this particular site.

Without this change, Composer throws an exception:
```
Project directory /app/ is not empty.
```

This commit also fleshes out the remaining steps for being able to
load the site locally.