Display media block elements on one line
The uswds_paragraphs module gives media block elements a width of
one half. Here I'm overriding that to give our media block elements
widths of one-third so our three elements will appear on one line.

Room for improvement/expansion includes dynamically calculating
the width based on the number of elements in the container -- as
this particular implementation still arbitrarily assumes the
number of elements that will be present -- or presenting width
settings in the UI.